ColorPickerPreference: Introduce set default button when attribute is set
Set a default value using the "defaultColorValue" settings namespace
attribute. The general idea is often we want set a value of say "-1"
and let framework handle it. One such scenario would be if you are
changing the color of a ImageView using ColorFilter. When framework
receives a "-1" value, framework sets the ImageView ColorFilter to null,
restoring the ImageView to original state.

Screenshot:

https://teameos.slack.com/files/bigrushdog/F09T5S155/screenshot_2015-08-31-04-02-54.png

Change-Id: I7962bad4a994712b7e511797b05a8160d1cc849e
Signed-off-by: xyyx <xyyx@mail.ru>
Signed-off-by: mohancm100 <mohanmanjappa@gmail.com>